Making it a bit more memory efficient

+++ SPJProject/src/quant_types.jl
@@ -8,8 +8,8 @@ end
 
 struct QuantMatrix{T, F} <: AbstractMatrix{T}
     matrix::Matrix{Chunk{T,F}}
-    dim::Pair{Int, Int}
-    blocksize::Int
+    dim::Pair{Int8, Int8}
+    blocksize::Int8
 end
 
 function Base.display(chunk::Chunk{T, F}) where T where F
